feat(*): update hook exports and rollup config
