+++ packages/hooks/package.json
@@ -5,11 +5,12 @@
     "sideEffects": false,
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
-    "module": "dist/esm/index.js",
+    "module": "dist/index.mjs",
     "files": [
         "dist"
     ],
     "scripts": {
+        "update:exports": "node ../../scripts/update-exports.js",
         "build": "yarn build:ts && yarn build:dts",
         "build:ts": "rollup -c",
         "build:dts": "dts-bundle-generator --no-banner --no-check --config dts.json"
@@ -34,5 +35,52 @@
     "publishConfig": {
         "access": "public"
     },
-    "gitHead": "10dd0711fab88c3173da9cd72abe50194559d021"
+    "gitHead": "10dd0711fab88c3173da9cd72abe50194559d021",
+    "exports": {
+        ".": {
+            "import": "./dist/index.mjs",
+            "require": "./dist/index.js"
+        },
+        "./package.json": "./package.json",
+        "./useClickOutside": {
+            "import": "./dist/useClickOutside/index.mjs",
+            "require": "./dist/useClickOutside/index.js"
+        },
+        "./useCountdown": {
+            "import": "./dist/useCountdown/index.mjs",
+            "require": "./dist/useCountdown/index.js"
+        },
+        "./useDidUpdateEffect": {
+            "import": "./dist/useDidUpdateEffect/index.mjs",
+            "require": "./dist/useDidUpdateEffect/index.js"
+        },
+        "./useEventCallback": {
+            "import": "./dist/useEventCallback/index.mjs",
+            "require": "./dist/useEventCallback/index.js"
+        },
+        "./useFocus": {
+            "import": "./dist/useFocus/index.mjs",
+            "require": "./dist/useFocus/index.js"
+        },
+        "./useImageLoadingState": {
+            "import": "./dist/useImageLoadingState/index.mjs",
+            "require": "./dist/useImageLoadingState/index.js"
+        },
+        "./useKeydownOutside": {
+            "import": "./dist/useKeydownOutside/index.mjs",
+            "require": "./dist/useKeydownOutside/index.js"
+        },
+        "./useMedia": {
+            "import": "./dist/useMedia/index.mjs",
+            "require": "./dist/useMedia/index.js"
+        },
+        "./usePrevious": {
+            "import": "./dist/usePrevious/index.mjs",
+            "require": "./dist/usePrevious/index.js"
+        },
+        "./useScrolledToBottomOfPage": {
+            "import": "./dist/useScrolledToBottomOfPage/index.mjs",
+            "require": "./dist/useScrolledToBottomOfPage/index.js"
+        }
+    }
 }
