--- packages/hooks/src/useClickOutside/hook.test.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { ReactTestRenderer } from 'react-test-renderer';
-import { renderHook } from '@testing-library/react-hooks';
-import { useClickOutside } from './hook';
-
-describe('Outside click', () => {
-    it('shuld work', () => {});
-});
