--- packages/hooks/src/usePrevious/hook.ts
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export function usePrevious<T>(value: T): T | undefined {
-    const ref = React.useRef<T>();
-
-    React.useEffect(() => {
-        ref.current = value;
-    });
-
-    return ref.current;
-}
