+++ packages/hooks/src/usePrevious/index.ts
@@ -1,1 +1,11 @@
-export * from './hook';
+import React from 'react';
+
+export function usePrevious<T>(value: T): T | undefined {
+    const ref = React.useRef<T>();
+
+    React.useEffect(() => {
+        ref.current = value;
+    });
+
+    return ref.current;
+}
